Trip Planner Application
This is a Trip Planner Application and will have the following features
in it:
--> The  user  will  be  able  to  edit  and  update  all  their
profile  contents  including  the
picture/avatar.
--> Each user will possess the following features:
o Discover people who are already registered.
o Send Friend requests.
o Accept Friend Requests
--> Users in a trip will be able to search places for visiting and can
add them to their trip.
--> Users in a trip will be able to delete the places from that trip.
--> Users in a trip will be able to see round-trip route on a google map
(embedded in an activity/fragment) with places, which are added already
to that trip.
--> Users in a trip will be able to launch Google Maps app(Android App
for Google Maps) to navigate between the places.